+++ sqln/driver/mysql/dialect.go
@@ -16,7 +16,7 @@ func (MySQLDialect) NotLike(field string, param string) string {
 	return fmt.Sprintf("%s NOT LIKE %s", field, param)
 }
 
-func (MySQLDialect) BuildPagination(query string, order string, limit uint16, offset uint16) string {
+func (MySQLDialect) BuildPagination(query string, order string, limit uint16, offset uint64) string {
 	return fmt.Sprintf("%s ORDER BY %s LIMIT %d OFFSET %d", query, order, limit, offset)
 }
