+++ sqln/driver/sqlserver/dialect.go
@@ -16,7 +16,7 @@ func (SQLServerDialect) NotLike(field string, param string) string {
 	return fmt.Sprintf("%s NOT LIKE %s", field, param)
 }
 
-func (SQLServerDialect) BuildPagination(query string, order string, limit uint16, offset uint16) string {
+func (SQLServerDialect) BuildPagination(query string, order string, limit uint16, offset uint64) string {
 	return fmt.Sprintf(
 		"%s ORDER BY %s OFFSET %d ROWS FETCH NEXT %d ROWS ONLY",
 		query, order, offset, limit,
